Move `StarterConfig` to `starter-data.ts`
