+++ src/@types/starter-data.ts
@@ -1,8 +1,9 @@
+import type PokemonSpecies from "#data/pokemon-species";
+import type { ElementalType } from "#enums/elemental-type";
 import type { MoveId } from "#enums/move-id";
+import type { Nature } from "#enums/nature";
 
-/**
- * Data for a single starter species
- */
+/** Data for a single starter species */
 export interface StarterDataEntry {
   moveset: StarterMoveset | StarterFormMoveData | null;
   eggMoves: number;
@@ -23,4 +24,16 @@ export type StarterMoveset = [MoveId] | [MoveId, MoveId] | [MoveId, MoveId, Move
 
 export interface StarterFormMoveData {
   [key: number]: StarterMoveset;
+}
+
+export interface StarterConfig {
+  species: PokemonSpecies;
+  dexAttr: bigint;
+  abilityIndex: number;
+  passive: boolean;
+  nature: Nature;
+  moveset?: StarterMoveset;
+  pokerus: boolean;
+  nickname?: string;
+  teraType?: ElementalType;
 }
