+++ src/ui/handlers/starter-select-ui-handler.ts
@@ -54,8 +54,7 @@ import { UiMode } from "#enums/ui-mode";
 import type { DexAttrProps, StarterAttributes, StarterPreferences } from "#system/game-data";
 import { DEFAULT_LANGUAGE_KEY } from "#system/supported-languages";
 import type { DexEntry } from "#types/dex-data";
-import type { StarterConfig } from "#types/starter-config";
-import type { StarterDataEntry, StarterMoveset } from "#types/starter-data";
+import type { StarterConfig, StarterDataEntry, StarterMoveset } from "#types/starter-data";
 import type { EnumValues } from "#types/utility-types";
 import type { ConfirmModeConfig } from "#ui/confirm-menu-config";
 import type { ConfirmUiHandler } from "#ui/confirm-ui-handler";
